+++ frontend/src/components/user/script.js
@@ -89,9 +89,7 @@ async function get_user() {
 		});
 	if (response.status === 200) {
 		const data = await response.json();
-
 		setCookie('user', JSON.stringify(data.user), 5 / 1440);
-
 		return data.user;
 	}
 	return null;
