+++ frontend/src/routes/account/page.html
@@ -10,7 +10,6 @@
 					<img src="../../static/assets/header/login.png" alt="logout image" class="buttons" id="logout-button">
 				</button>
 			</div>
-			<div id="matchList">Match History</div>
 			<div id="sheet">
 				<img src="../routes/assets/images/account/profile_sheet_3.png" alt="sheet image" id="sheet-image">
 				<div id="main-page">
@@ -61,12 +60,6 @@
 							</label>
 							<input type="file" id="edit-profile-picture" name="edit-profile-picture" accept="image/*">
 						</form>
-					</div>
-					<div  class="stats text" id="wins">
-						<span>WINS</span>
-					</div>
-					<div  class="stats text" id="losses">
-					<span>LOSSES</span>
 					</div>
 					<div id="delete-profile">
 						<button class="buttons text" >
