+++ frontend/src/routes/account/style.css
@@ -53,50 +53,6 @@ body #page #background #container .buttons {
 body #page #background #container .buttons:hover {
   transform: scale(1.2);
 }
-body #page #background #container #matchList {
-  font-size: 0.7rem;
-  position: absolute;
-  left: 100px;
-  font-family: isaac-bold;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: auto;
-  max-height: 300px;
-  overflow-y: scroll;
-}
-body #page #background #container #matchList::-webkit-scrollbar {
-  display: none;
-}
-body #page #background #container #matchList .Matchline {
-  position: relative;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  gap: 10px;
-  margin-top: 10px;
-  margin-bottom: 10px;
-}
-body #page #background #container #matchList .Matchline .player-left {
-  text-align: left;
-  width: 90px;
-  height: 20px;
-  white-space: nowrap;
-  overflow: hidden;
-  padding-right: 20px;
-  display: flex;
-  align-items: center;
-}
-body #page #background #container #matchList .Matchline .player-right {
-  text-align: right;
-  width: 90px;
-  height: 20px;
-  white-space: nowrap;
-  overflow: hidden;
-  padding-left: 20px;
-  display: flex;
-  align-items: center;
-}
 body #page #background #container #logout-sheet {
   width: 19%;
   position: absolute;
@@ -129,7 +85,7 @@ body #page #background #container #sheet {
   position: relative;
 }
 body #page #background #container #sheet #sheet-image {
-  width: 50vw;
+  width: 47vw;
   display: flex;
   justify-content: center;
   align-items: center;
