+++ frontend/src/routes/friends/page.html
@@ -1,7 +1,6 @@
 {{header}}
 <div id="page">
-	<div class="background">
-		
+	<div id="background">
 	</div>
 </div>
