+++ frontend/src/routes/stats/page.html
@@ -1,0 +1,37 @@
+{{header}}
+<div id="page">
+	<div id="background">
+		<div id="book">
+			<img src="../routes/assets/images/stats/book.png" alt="Book">
+			<div id="left-page">
+				<span class="text">Match History</span>
+				<div id="matchList">
+
+				</div>
+			</div>
+			<div id="right-page">
+				<span class="text"> Stats</span>
+				<div id="pong-stats">
+					<div class="stats text" id="wins">
+						<span>WINS</span>
+						<img src="../routes/assets/images/stats/wins-logo.png" alt="win Logo">
+						<div id="wins-count" class="stats-count">
+						</div>
+					</div>
+					<div  class="stats text" id="losses">
+						<span>LOSSES</span>
+						<img src="../routes/assets/images/stats/losses-logo.png" alt="loose Logo">
+						<div id="losses-count" class="stats-count">
+						</div>
+					</div>
+					<div id="flappy-stats">
+
+					</div>
+				</div>
+			</div>
+		</div>
+	</div>
+</div>
+	
+<script src="js/index.js"></script>
+<script src="https://accounts.google.com/gsi/client" async defer></script>
