+++ frontend/src/routes/stats/style.sass
@@ -1,0 +1,180 @@
+@font-face
+  font-family: isaac-bold
+  src: url('../assets/font/isaac_bold.ttf')
+
+body, html
+    background-image: url('../assets/images/account/background_account.png')
+    background-position: center
+    background-repeat: no-repeat
+    background-size: cover
+    height: 100%
+
+    #page
+        // position: relative
+        height: 100%
+
+        .text
+            font-family: isaac-bold
+            color: #332C2B
+        button
+            background-color: transparent
+            border: none
+       
+        button:hover
+            transform: scale(1.2)
+
+        #background
+            position: relative
+            background-position: center
+            background-repeat: no-repeat
+            background-size: contain
+            display: flex
+
+            #book
+                position: relative
+                display: flex
+                justify-content: center
+                width: 100%
+                font-family: isaac-bold
+                
+                img
+                    width: 70vw
+                
+                #left-page
+                    // background-color: #cf0000
+                    position: absolute
+                    top: 8%
+                    left: 19%
+                    width: 30%
+                    height: 80%
+                    display: flex
+                    justify-content: center
+
+                    span
+                        font-size: 2vw
+
+                    #matchList 
+                        position: absolute
+                        top: 10%
+                        left: 0%
+                        display: flex
+                        flex-direction: column
+                        align-items: center
+                        width: 100%
+                        overflow-y: scroll
+                        font-size: 1.2vw
+
+                        &::-webkit-scrollbar
+                            display: none
+
+                        .Matchline 
+                            position: relative
+                            display: flex
+                            flex-direction: row
+                            justify-content: space-between
+                            gap: 10px
+                            margin-top: 10px
+                            margin-bottom: 10px
+
+                            .player-left
+                                overflow-wrap: break-word
+                                text-align: left
+                                white-space: nowrap
+                                overflow: hidden
+                                padding-right: 20px
+                                display: flex
+                                align-items: center
+
+                            .player-right
+                                overflow-wrap: break-word
+                                text-align: right
+                                white-space: nowrap
+                                overflow: hidden
+                                padding-left: 20px
+                                display: flex
+                                align-items: center
+
+                #right-page
+                    // background-color: #0400fd
+                    position: absolute
+                    top: 8%
+                    left: 51%
+                    width: 30%
+                    height: 80%
+                    display: flex
+                    justify-content: center
+
+                    span
+                        font-size: 2vw
+                    
+                    #pong-stats
+                    
+                        span
+                            font-size: 1.5vw
+                        
+                        
+                        .stats-count
+                                // background-color: #00ff00
+                                width: 100%
+                                height: 100%
+                                display: flex
+                                justify-content: center
+                                font-size: 3vw
+
+                        #wins
+                            // background-color: #00ff00
+                            position: absolute
+                            top: 10%
+                            left: 3%
+                            height: 43%
+                            width: 44%
+                            display: flex
+                            justify-content: flex-start
+                            align-items: center
+                            flex-direction: column
+
+                            img
+                                padding-bottom: 4%
+                                height: 18%
+                                width: 30%
+
+                        #losses
+                            // background-color: #ff0000
+                            position: absolute
+                            top: 10%
+                            left: 52%
+                            height: 43%
+                            width: 44%
+                            display: flex
+                            justify-content: flex-start
+                            align-items: center
+                            flex-direction: column
+
+                            img
+                                rotate: -6deg
+                                width: 30%
+                                height: 21%
+
+
+                        &::-webkit-scrollbar
+                            display: none
+.win 
+    background-image: url(../../static/assets/yes.png)
+    width: 20px
+    height: 20px
+    background-size: contain
+    background-repeat: no-repeat
+    background-position: center
+    background-color: transparent
+    outline: none
+    border: none
+.lost 
+    background-image: url(../../static/assets/no.png)
+    width: 20px
+    height: 20px
+    background-size: contain
+    background-repeat: no-repeat
+    background-position: center
+    background-color: transparent
+    outline: none
+    border: none
